+++ visualize/src/main.c
@@ -25,7 +25,7 @@ void			map_free(t_matrix *matrix)
 char			**alocate_matrix(size_t height, size_t width)
 {
 	char		**matrix;
-	int			j;
+	size_t		j;
 
 	j = 0;
 	if (!(matrix = (char **)malloc(sizeof(char *) * height + 1)))
@@ -42,9 +42,9 @@ char			**alocate_matrix(size_t height, size_t width)
 
 char			**reading_create_map(char **matrix, char *line, t_matrix *size)
 {
-	int			x;
-	int			y;
-	int			i;
+	size_t		x;
+	size_t		y;
+	size_t		i;
 
 	x = 0;
 	while (x < size->size_map_x && get_next_line(0, &line))
@@ -66,8 +66,8 @@ char			**reading_create_map(char **matrix, char *line, t_matrix *size)
 
 void			print_map(char **map, t_player *player)
 {
-	int			i;
-	int			j;
+	size_t		i;
+	size_t		j;
 	char		filler;
 
 	i = 0;
